Move json serializer into subpackage
