+++ src/sciline/serialize/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+
+from ._json import json_serialize_task_graph
+
+__all__ = ['json_serialize_task_graph']
